adding mypy for type checking
This adds mypy and isort for checking pystemd
