+++ pystemd/cutils.pyi
@@ -1,0 +1,1 @@
+def setns(fd: int, nstype: int) -> int: ...
