+++ pystemd/systemd1/__init__.py
@@ -9,5 +9,4 @@
 from .manager import Manager
 from .unit import Unit
 
-
 __all__ = ["Manager", "Unit"]
